Resolve "Refactor sensor accuracy"

Closes #53

See merge request se-anwendungsprojekt-22-23/sensing-plugin!34

+++ ios/Classes/Sensors/ISensorStreamHandler.swift
@@ -32,7 +32,7 @@ protocol ISensorStreamHandler: NSObject, FlutterStreamHandler {
         - timeInterval: required interval in ms
      - Returns: ResultWrapper if the time interval was changed sucessfully, otherwise StateIndicator.failure
      */
-    func changeSensorTimeInterval(timeInterval: Int32) -> ResultWrapper
+    func changeSensorTimeInterval(timeInterval: Int64) -> ResultWrapper
 
     /**
      Get the information about this sensor
